pyproject.toml
[tool.poetry]
name = "assistant"
version = "0.0.0"
description = "Assistant"
authors = ["Stefans Mezulis <stefans.mezulis@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/stefansm/assistant"
repository = "https://github.com/stefansm/assistant"
documentation = "https://assistant.readthedocs.io"
classifiers = [
    "Development Status :: 1 - Planning",
]

[tool.poetry.urls]
Changelog = "https://github.com/stefansm/assistant/releases"

[tool.poetry.dependencies]
python = "^3.10"
click = ">=8.0.1"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
tiktoken = "^0.4.0"

[tool.poetry.dev-dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
flake8 = ">=4.0.1"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}

[tool.poetry.scripts]
assistant = "assistant.__main__:main"

[tool.poetry.group.dev.dependencies]
types-click = "^7.1.8"
flake8-bugbear = "^23.7.10"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["assistant", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_context = true

[[tool.mypy.overrides]]
module = "openai"
ignore_missing_imports = true
disallow_untyped_calls = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
